share user profile data between login and header components
